Standardize product tax: decimal rate, line-level rounding, per-line persistence
- shop_tax.value: integer -> decimal(8,4) to support fractional tax rates
- Add gp247_line_tax() helper: single source of line-level tax rounding
- ShopCurrency::sumCartCheckout(): use line-level rounding (was per-unit)
- ShopOrder::createOrder(): compute shop_order_detail.tax with the shared
  helper on the options-included, currency-converted base so
  shop_order.tax == Sum(shop_order_detail.tax)
- ShopTax: cast value to float

Modification 20260803T223543.

+++ src/Admin/Database/Migrations/00_00_00_create_tables_shop.php
@@ -390,7 +390,10 @@ return new class extends Migration
             function (Blueprint $table) {
                 $table->increments('id');
                 $table->string('name', 255);
-                $table->integer('value')->default(0);
+                // WHY: decimal (not integer) so tax rates can carry fractional percents
+                // such as 8.5% or the US 8.375% (modification 20260803T223543, ADR
+                // shop-admin_tax-standardization, decision D1). 8,4 = up to 9999.9999%.
+                $table->decimal('value', 8, 4)->default(0);
                 $table->timestamps();
             }
         );
